Update Dockerfile, .dockerignore, mainwindow.cpp, displayRaw.cpp and CMakeLists.txt
In the Dockerfile, 'git' was added as an installation requirement and 'cmake' parameters were refactored. The '.jpeq' and '.jpg' ignore files were removed from .dockerignore. The deprecated 'boost/bind.hpp' was updated into the mainwindow.cpp. In the displayRaw.cpp, the method for retrieving screen size was updated for better reliability. The minimum required version of cmake was changed to 3.5 for better compatibility in the CMakeLists.txt.