Added function to list available camera models and improved UI updates
A new function GetAvailableCameraModels() was added to mainwindow class which lists out all available camera models. This is an important feature to allow the application to support multiple types of cameras.

In addition, certain UI changes were refactored to be triggered from the main thread using QMetaObject::invokeMethod. This helps in avoiding potential synchronization issues and improve overall execution because Qt's GUI routines are not thread-safe and need to be called from the main thread.