Add camera selection feature and UI enhancements
Updated the code to include camera selection feature enabling the use of multiple types of cameras through the addition of the GetAvailableCameraModels() function. This function fetches all the models of the cameras available for use improving usability of the system. Certain UI calls were moved to the main thread using QMetaObject::invokeMethod.