Update code documentation and remove unused file
Removed default_defines.h as it was unnecessary and not in use.
Updated copyright and author information in all headers.
Minor code changes were made to ensure value safety, remove unused inclusions and to use modern C++ idioms.
CMake was updated to reflect the removed file. The LICENSE.md file was also added.