Refactor project for improved testing and extend functionality
Implemented a wrapper for the XIMEA API to create a clear boundary and to enable unit testing of the applications' logic. A mock implementation of the API was created for testing. Additionally, removed unnecessary tests and wrote new ones covering more of the application. Refactored file names and methods to follow common naming conventions. Also, small changes were made in the Dockerfile and gitlab-CI yaml to enable GUI related tests in the CI using xvfb.